+++ main.cpp
@@ -3,79 +3,12 @@
 #include "my_operator.h"
 #include "my_vector.h"
 #include "my_reference.h"
+#include "my_list.h"
 using namespace std;
 
-//简单的list
-template<typename T>
-//节点
-struct list_node {
-    list_node<T>* prev;
-    list_node<T>* next;
-    T data;
-};
-//list
-template<typename T>
-struct my_list {
-    typedef  list_node<T>* iterator;
-    my_list() {
-        node==new list_node<T>;
-        node->next=node;
-        node->prev=node;
-    }
-    iterator begin() {
-        return node->next;
-    }
-    iterator end() {
-        return node;
-    }
-
-    void insert(iterator pos,const T& value) {
-        iterator tmp=new list_node<T>;
-        tmp->data=value;
-        tmp->next=pos->next;
-        tmp->prev=pos;
-        pos->next->prev=tmp;
-        pos->next=tmp;
-    }
-
-    void erase(iterator pos) {
-        pos->next->prev=pos->prev;
-        pos->prev->next=pos->next;
-        delete pos;
-    }
-
-    void push_back(const T& value) {
-        insert(node,value);
-    }
-
-    iterator pop_back() {
-        iterator tmp=node->prev;
-        erase(node->prev);
-        return tmp;
-    }
-private:
-    iterator node;//链表的结尾
-};
-
-template<typename T>
-void  printVec(T begin,T end) {
-    while(begin!=end) {
-        cout<<*begin<<" ";
-        begin++;
-    }
-}
-
-template<typename T>
-void  printList(list_node<T>* begin,list_node<T>* end) {
-    while(begin!=end) {
-        cout<<begin->data<<" ";
-        begin=begin->next;
-    }
-}
-
 int main() {
-    //int  a[5]={1,2,3,4,5};
-    //printVec(a,a+5);
+    int  a[5]={1,2,3,4,5};
+    printVec(a,a+5);
     my_list<int>  list;
     list.push_back(1);
     list.push_back(2);
